chore(tests): Move tests/snuba/api/endpoint tests to tests/sentry
Without this change, it could sometimes be hard to find the tests since they would not match the mental mapping of `src/sentry` to `tests/sentry`.
