+++ .github/CODEOWNERS
@@ -214,17 +214,53 @@ yarn.lock                                                @getsentry/owners-js-de
 /src/sentry/api/endpoints/organization_events_stats.py                      @getsentry/visibility
 /src/sentry/api/endpoints/organization_measurements_meta.py                 @getsentry/visibility
 
-/tests/snuba/api/endpoints/*                                                @getsentry/visibility
-/tests/snuba/api/endpoints/test_organization_tags.py                        @getsentry/visibility
-/tests/snuba/api/endpoints/test_organization_events_histogram.py            @getsentry/visibility
-/tests/snuba/api/endpoints/test_organization_events_facets_performance.py   @getsentry/visibility
-/tests/snuba/api/endpoints/test_organization_events_spans_performance.py    @getsentry/visibility
-/tests/snuba/api/endpoints/test_organization_events_spans_histogram.py      @getsentry/visibility
-/tests/snuba/api/endpoints/test_organization_events_trace.py                @getsentry/visibility
-/tests/snuba/api/endpoints/test_organization_events_trends.py               @getsentry/visibility
-/tests/snuba/api/endpoints/test_organization_events_trends_v2.py            @getsentry/visibility
-/tests/snuba/api/endpoints/test_organization_events_vitals.py               @getsentry/visibility
-/tests/snuba/api/endpoints/test_organization_tagkey_values.py               @getsentry/visibility
+/tests/sentry/api/endpoints/test_discover_homepage_query.py                                    @getsentry/visibility
+/tests/sentry/api/endpoints/test_discover_key_transactions.py                                  @getsentry/visibility
+/tests/sentry/api/endpoints/test_discover_saved_queries.py                                     @getsentry/visibility
+/tests/sentry/api/endpoints/test_discover_saved_query_detail.py                                @getsentry/visibility
+/tests/sentry/api/endpoints/test_group_details.py                                              @getsentry/visibility
+/tests/sentry/api/endpoints/test_group_event_details.py                                        @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_event_details.py                                 @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_eventid.py                                       @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events.py                                        @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_facets.py                                 @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_facets_performance.py                     @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_facets_performance_histogram.py           @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_has_measurements.py                       @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_histogram.py                              @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_mep.py                                    @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_meta.py                                   @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_profile_functions_metrics.py              @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_span_indexed.py                           @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_span_metrics.py                           @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_span_ops.py                               @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_spans_histogram.py                        @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_spans_performance.py                      @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_stats.py                                  @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_stats_mep.py                              @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_stats_profile_functions_metrics.py        @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_stats_span_indexed.py                     @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_stats_span_metrics.py                     @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_trace.py                                  @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_trends.py                                 @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_trends_v2.py                              @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_events_vitals.py                                 @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_group_index_stats.py                             @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_issues_resolved_in_release.py                    @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_measurements_meta.py                             @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_metrics_meta.py                                  @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_sessions.py                                      @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_spans_aggregation.py                             @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_spans_trace.py                                   @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_stats_summary.py                                 @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_stats_v2.py                                      @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_tagkey_values.py                                 @getsentry/visibility
+/tests/sentry/api/endpoints/test_organization_tags.py                                          @getsentry/visibility
+/tests/sentry/api/endpoints/test_project_event_details.py                                      @getsentry/visibility
+/tests/sentry/api/endpoints/test_project_events.py                                             @getsentry/visibility
+/tests/sentry/api/endpoints/test_project_group_index.py                                        @getsentry/visibility
+/tests/sentry/api/endpoints/test_project_tags.py                                               @getsentry/visibility
+
 
 /src/sentry/api/endpoints/organization_transaction_anomaly_detection.py     @getsentry/data
