+++ tests/sentry/api/endpoints/test_discover_homepage_query.py
