+++ tests/sentry/api/endpoints/test_discover_saved_query_detail.py
