+++ tests/sentry/api/endpoints/test_project_event_details.py
