+++ tests/sentry/api/endpoints/test_project_tags.py
